factored out histfactory factories
